+++ OnlineShopV1/Core/Models/Product.cs
@@ -1,5 +1,7 @@
 using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.Threading.Tasks;
 
 namespace OnlineShopV1
 {
@@ -10,5 +12,14 @@ namespace OnlineShopV1
         
         [Required]    
         public String Name { get; set; }
+    }
+    
+    
+    public interface IProductRepository
+    {
+
+        Task<List<Product>> ListProducts();
+        Task<Product> GetByIdAsync(int id);
+        Task AddAsync(Product product);
     }
 }
